requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Admin==1.6.1
Flask-BabelEx==0.9.4
Flask-BasicAuth==0.2.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
python-decouple==3.8
pytz==2023.2
requests==2.28.2
speaklater==1.3
SQLAlchemy==2.0.7
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
yarl==1.8.2
